The Eight,
finished up the first version of the Eight. Basic cost as regular Units with regular equipment.
Special Rules not priced in, Special Weapons Priced for 1 PL per Weapon (Affects the 2 Melee Weapons) and Shield Generators (Missing in Apocalypse) for 1 PL Each (comparable to a Shield Drone).
You may not change the base Equipment but can chosse drones at the moment.
Detachement Rule to have Farsight as the Commander and grant the 24" range from Commander from the Infiltrators Rule.

Might have to revisit costs and wording on some of the new special rules.