Transports.
Redone the Transport Limits, updated them to mostly the 10th edition versions of the number of models.
This means you can usually transport a character and a squad.

Additionally Space Marines got the restrictions to Primaris and First Born removed. This was mostly done with the bigger transports in 10th, but not for everything.
We could not care less if a Intercessor rides in a Rhino, we live in a world where adapter cables exist.