Ruinstorm Daemons
Finished the first version of the Ruinstorm Daemons.
Ruinstorm Daemons are Horus Heresys version of Different kind of Daemons, kind of Daemons Undivided, Rules to repesent different Versions of Daemons that aren't part of a specific god.
Created based on a combination of what rules they got in Horus Heresy and adjusted to the baseline of what the existing Daemons offer.