Added Missing Categories to Units
Knight Ruinator Profile is out - changes to the Terrorpulse Missile Launcher.
Buffed Fellbore to not be balanced vs. other Knight melee weapons, but a bit better vs vehicles.
Darkflame Lance kept the same - feels allright for now.
Removed the Carapace Weapons, is not an Option for this Knight.
Has a Headhunter Special Rule and one that buffs Wardogs around itself.
Lets keep it without one for the moment - maybe add it when it has more of an identity later.