Added some Experimental Units to the game
Increased the Cost for the Endless Phalanx unit to 160/270/390 PL
Enabled the Unit as a LoW Choice.
(100-200 Necron Warriors as one Unit. It has 1 ranged and 1 melee weapon per 50 warriors, to reduce "Split Fire" - better Reanimation Protocols, could be killed easily if opponent stacks shit-tons of Blast marker on them)

Added the Plaguereaper as a CSM / Death Guard LoW Unit
Baneblade with a giant SAP Flamer and Nurlgings for extra Melee punch.

Remodeled the Green Tide to closer resemble the Version of Apocalypse Reloaded with the Mega-Orkboymob.
Added an increase in PL Cost depending on the start-Size of the Unit. (needs testing to closer resemble our own rule of bigger units having higher costs the more Wounds they have)

Added Tower of Skulls as a CSM / World Eater LoW Unit.
Baneblade with different Weapons that can regenerate Wounds.

Added Van Staab as a HQ Choice to Orks - a Warzone Armageddon Named character.