Agents of Imperium changes
Sister of Silence:

Moved to Adeptus Custodes. Are missing a shared Faction Keyword with Custodes to do a combined force similiar to the Cult Legions.
Added the Sister HQ. (Reroll Wound Aura for SoS)

Arbites:

Added to Agents of the Imperium with their 3 Unit types

Navis Imperialis:

Added Imperial Navy Breacher to Agents of the Imperium

Combat Assets:
Changed how the primary tactical assets are selected, in a similiar way on how Agents have all choices preselected.

Assassins:
Removed the Infiltrators Rule from Assassins.
Added no Out of Command to the Independed Operative (cannot be a Warlord) ability.

Added a LoW choice, to take the 4 different assassins as one choice. If taken this way, they are granted the ability to receive their own orders.