More Ultramrine Bling
Added the new Version of Cato Sicarius as an option. Trading his Captain rank for a Wound Re-Roll, a stronger Aura and +1 to wound.

Added the Banner of Macragge to one Vitrix Guard Unit in your army - +1 Attack in Melee Aura for Ultramarine Light units.

Added Uriel Ventris - Captain with a better Blade and Master of the Fleet rank - grants himself and 3 other infantry units in his detachment Deepstrike.